backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-magic>=0.4.27
mail-parser>=3.15.0
extract-msg>=0.48.0
lark>=1.1.0
imgkit>=1.2.3
chardet>=5.0.0
python-dateutil>=2.8.0
yara-python>=4.3.0
